[ld2420] Rename c++ files for predictable doxygen generation
