+++ esphome/components/ld2420/text_sensor/ld2420_text_sensor.h
